Move tile related stuff from 'types' to 'tile'
