+++ src/js/tiles/concordance/model.ts
@@ -22,10 +22,10 @@ import { Action, ActionDispatcher, SEDispatcher, StatelessModel } from 'kombo';
 import { Observable } from 'rxjs';
 import { concatMap } from 'rxjs/operators';
 import { AppServices } from '../../appServices';
-import { PCRequestArgs, RequestArgs } from '../../common/api/kontext/concordance';
 import { Line, ConcResponse } from '../../common/api/abstract/concordance';
 import { ConcordanceMinState, IStateArgsMapper } from '../../common/models/concordance';
-import { Backlink, BacklinkWithArgs, HTTPMethod, SystemMessageType, isSubqueryPayload, DataApi } from '../../common/types';
+import { HTTPMethod, SystemMessageType, isSubqueryPayload, DataApi } from '../../common/types';
+import { Backlink, BacklinkWithArgs } from '../../common/tile';
 import { ActionName as GlobalActionName, Actions as GlobalActions } from '../../models/actions';
 import { findCurrLemmaVariant, WdglanceMainFormModel } from '../../models/query';
 import { importMessageType } from '../../notifications';
