+++ src/js/tiles/empty.ts
@@ -15,7 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ITileProvider, QueryType, TileComponent } from '../common/types';
+import { QueryType } from '../common/types';
+import { ITileProvider, TileComponent } from '../common/tile';
 
 
 /**
