+++ src/js/tiles/freqPie/index.ts
@@ -20,7 +20,8 @@ import * as Immutable from 'immutable';
 import { AppServices } from '../../appServices';
 import { MultiBlockFreqDistribAPI, FreqSort } from '../../common/api/kontext/freqs';
 import { FreqDataBlock, SubqueryModeConf } from '../../common/models/freq';
-import { ITileProvider, LocalizedConfMsg, QueryType, TileComponent, TileConf, TileFactory } from '../../common/types';
+import { LocalizedConfMsg, QueryType } from '../../common/types';
+import { ITileProvider, TileComponent, TileConf, TileFactory } from '../../common/tile';
 import { puid } from '../../common/util';
 import { factory as defaultModelFactory, FreqPieDataRow, FreqPieModel } from './model';
 import { factory as subqModelFactory } from './subqModel';
