+++ src/js/tiles/geoAreas/index.ts
@@ -20,7 +20,8 @@ import { ActionDispatcher } from 'kombo';
 
 import { AppServices } from '../../appServices';
 import { DataRow, FreqDistribAPI, FreqSort } from '../../common/api/kontext/freqs';
-import { ITileProvider, QueryType, TileComponent, TileConf, TileFactory } from '../../common/types';
+import { QueryType } from '../../common/types';
+import { ITileProvider, TileComponent, TileConf, TileFactory } from '../../common/tile';
 import { GeoAreasModel } from './model';
 import { init as viewInit } from './views';
