+++ src/js/tiles/mergeCorpFreq/model.ts
@@ -23,7 +23,8 @@ import { map } from 'rxjs/operators';
 import { AppServices } from '../../appServices';
 import { APIResponse, BacklinkArgs, DataRow, FreqDistribAPI, SingleCritQueryArgs } from '../../common/api/kontext/freqs';
 import { callWithRequestId } from '../../common/api/util';
-import { Backlink, BacklinkWithArgs, HTTPMethod } from '../../common/types';
+import { HTTPMethod } from '../../common/types';
+import { Backlink, BacklinkWithArgs } from '../../common/tile';
 import { ActionName as GlobalActionName, Actions as GlobalActions } from '../../models/actions';
 import { ConcLoadedPayload } from '../concordance/actions';
 import { DataLoadedPayload } from './actions';
