+++ src/js/tiles/treq/index.ts
@@ -18,7 +18,8 @@
 import * as Immutable from 'immutable';
 
 import { AppServices } from '../../appServices';
-import { ITileProvider, QueryType, TileComponent, TileConf, TileFactory } from '../../common/types';
+import { QueryType } from '../../common/types';
+import { ITileProvider, TileComponent, TileConf, TileFactory } from '../../common/tile';
 import { SearchPackages, TreqAPI, TreqTranslation } from '../../common/api/treq';
 import { TreqModel } from './model';
 import { init as viewInit } from './view';
