+++ src/js/tiles/treqSubsets/index.ts
@@ -17,7 +17,8 @@
  */
 
 import * as Immutable from 'immutable';
-import { ITileProvider, TileComponent, QueryType, TileConf, TileFactory } from '../../common/types';
+import { QueryType } from '../../common/types';
+import { ITileProvider, TileComponent, TileConf, TileFactory } from '../../common/tile';
 import { TreqSubsetModel, TranslationSubset } from './model';
 import { TreqAPI, TreqTranslation } from '../../common/api/treq';
 import {init as viewInit} from './view';
