+++ src/js/tiles/wordFreq/views/index.tsx
@@ -19,7 +19,7 @@ import * as Immutable from 'immutable';
 import { ActionDispatcher, BoundWithProps, ViewUtils } from 'kombo';
 import * as React from 'react';
 
-import { CoreTileComponentProps, TileComponent } from '../../../common/types';
+import { CoreTileComponentProps, TileComponent } from '../../../common/tile';
 import { ActionName as GlobalActionName, Actions as GlobalActions } from '../../../models/actions';
 import { GlobalComponents } from '../../../views/global';
 import { SummaryModel, SummaryModelState } from '../model';
